requirements.txt
captum==0.8.0
matplotlib==3.10.3
mne==1.9.0
munch==4.0.0
numpy==2.3.1
pandas==2.3.0
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
torch==2.2.0
tqdm==4.66.4
transformers==4.53.0
